workers/SharedWorker-constructor.html

<!DOCTYPE html>
<title>Test SharedWorker constructor functionality.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

async_test(function(t) {
  try {
    new SharedWorker({toString:function(){throw "exception"}}, "name")
    assert_unreached("FAIL: toString exception not propagated.");
  } catch (ex) {
      assert_equals(ex, "exception", "PASS: toString exception propagated correctly.")
      if (ex != "exception")
        assert_unreached("FAIL: unexpected exception (" + ex + ") received instead of one propagated from toString.");
      t.done();
  }
}, "Test toString exception propagated correctly.");

async_test(function(t) {
  assert_throws(new RangeError("RangeError: Maximum call stack size exceeded"),
                function() {
                    var foo = {toString:function(){new Worker(foo)}}
                    new SharedWorker(foo, name); },
                "PASS: trying to create workers recursively resulted in an exception.");
  t.done();
}, "Test recursive worker creation results in exception.");

async_test(function(t) {
  assert_throws(new TypeError("TypeError: Failed to construct 'SharedWorker': 1 argument required, but only 0 present."),
                function() { new SharedWorker(); },
                "PASS: invoking SharedWorker constructor without arguments resulted in an exception.");
  t.done();
}, "Test SharedWorker creation without arguments results in exception.");

async_test(function(t) {
  try {
    var worker = new SharedWorker("support/SharedWorkerCommon.js");
  } catch (ex) {
    assert_unreached("FAIL: Constructor failed when no name is passed: (" + ex + ")");
  }
  t.done();
}, "Test SharedWorker constructor without a name does not result in an exception.");

async_test(function(t) {
  try {
    var worker = new SharedWorker("support/SharedWorkerCommon.js", null);
  } catch (ex) {
    assert_unreached("FAIL: Constructor failed when null name is passed: (" + ex + ")");
  }
  t.done();
}, "Test SharedWorker constructor with null name does not result in an exception.");

async_test(function(t) {
  try {
    var worker = new SharedWorker("support/SharedWorkerCommon.js", undefined);
  } catch (ex) {
    assert_unreached("FAIL: Constructor failed when undefined name is passed: (" + ex + ")");
  }
  t.done();
}, "Test SharedWorker constructor with undefined name does not result in an exception.");

async_test(function(t) {
  try {
    var worker = new SharedWorker("support/SharedWorkerCommon.js", "name");
  } catch (ex) {
    assert_unreached("FAIL: invoking SharedWorker constructor resulted in an exception: (" + ex + ")");
  }
  t.done();
}, "Test SharedWorker constructor suceeds.");

</script>
